Moved check-time dependencies that are not actually used into separate folder
Looks like now Chef started to validate dependencies before the scripts start to run.
Therefore some cookboos that are declared (but not used) cause errors because they don't
exist. There were 2 choices - either to manually remove unnecessary dependencies from
every cookbook (and complicate further updates of cookbooks) or to put them into the
folder with other cookbooks. The latter was chosen as it requires less manual changes.

For cookbooks that are not actually used but are declared as dependencies a new folder
was created: `cookbooks-dependencies` so that we don't mash everything up.